# Conflicts:
#	src/main/java/com/example/sbt_final_hr/domain/service/EmployeesService.java

+++ src/main/java/com/example/sbt_final_hr/domain/model/entity/EmployeesSkill.java
@@ -34,8 +34,11 @@ public class EmployeesSkill {
     private Skills skill;
     //chatgpt는 skillLanguage 말고 skill 이렇게 쓰라는데 모르겠네. DB컬럼명과 혼동된다고해서.
     //skillLanguage->skill
+
+    //8월 3일 21:23 소수점 값을 더해주기 위해서 int -> double 변경하였습니다.
     @Column(name = "Skill_Career", nullable = false)
-    private int skillCareer;
+    private double skillCareer;
+
 //
 //    @OneToMany(mappedBy = "employee", cascade = CascadeType.ALL)
 //    private List<EmployeesSkill> skills;
